Make `Object` a nil-able type
This makes almost every `Object` type into `T?`, and allows for the
convenient creation of a null Object, which was previously impossible.

Requires the usage of type refinement, e.g. `assert`, to ensure an
`Object` is non-nil before using it. A new function, `nonnil`, has also
been introduced for this. Doing this is technically more type safe, but
also more verbose in many situations, especially with chained method calls.

In the future, this may change (e.g. determining whether methods actually
return null values or not).