Remove `__namecall` on script instances
It became pretty clear that this is not what Luau intended `__namecall`
to be used for.

Calling from `__namecall` back into Lua has been repeated cause for
trouble (e.g. with yielding). While it isn't needed for debugging
(`lua_break` is not strictly necessary) I think it is best to remove
this functionality.

For performance, `__namecall` is retained for instances with no script
attached. That falls under the intended use case (C calls on C userdata).

- "Quirk" of not being able to yield in a method call is now resolved.
- Default arguments are no longer "free" (they must be registered with
  the engine and handled on the Luau side too, for example with `?`
  and the `or` operator).