Remove the `@classType` annotation
In the vast majority of cases, the `@classType` annotation did not carry any
useful information. Typically, and now as a requirement, the type is the same
name as the class's `local`. The type name is now considered reserved by the
`@class` annotation.

Existing `@classType` annotations will be ignored as types no longer allow any
annotations. Class types should be renamed to have the same name as the class
local, if they aren't already.