Make everything nullable
Maintaining a list of methods with non-nullable return values (usually
by inspecting the Godot source) is not sustainable. Using `assert` will
be annoying but is a decent workaround and leads to a safer (a bit) type
system.

FWIW, this model is adopted by Godot Rust as well (i.e., all returned
objects are `Option`).